requirements.txt
filelock~=3.13.1
aiogram~=3.16.0
python-dotenv~=0.21.0